code/requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
ci-info==0.3.0
click==8.1.3
configobj==5.0.6
configparser==5.3.0
easyocr==1.6.2
etelemetry==0.3.0
filelock==3.8.0
future==0.18.2
httplib2==0.20.4
idna==3.4
imageio==2.22.0
isodate==0.6.1
looseversion==1.0.1
lxml==4.9.1
networkx==2.8.6
nibabel==4.0.2
ninja==1.10.2.4
nipype==1.8.5
numpy==1.23.3
opencv-python-headless==4.5.4.60
packaging==21.3
pandas==1.5.0
pathlib==1.0.1
Pillow==9.2.0
prov==2.0.0
pyclipper==1.3.0.post3
pydot==1.4.2
PyMuPDF==1.20.2
pyparsing==3.0.9
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
pyxnat==1.5
PyYAML==6.0
rdflib==6.2.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.1
Shapely==1.8.4
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
tabulate==0.8.10
tifffile==2022.8.12
torch==1.12.1
torchvision==0.13.1
traits==6.3.2
typing_extensions==4.3.0
urllib3==1.26.12
